chore: refactor duplicated types into shared lib
Move four sets of types that were defined separately in both client and
server into @mini-infra/types so there is a single source of truth:

- SelfUpdateState, SelfUpdateStatus, SelfUpdateCheckResult → new lib/types/self-update.ts
- StackAdoptionCandidate, StackAdoptionCandidatesResponse → lib/types/stacks.ts
  (resolves naming collision with the HAProxy-context EligibleContainer in deployments.ts)
- TestDockerRegistryRequest, TestDockerRegistryResponse → lib/types/settings.ts
- PortValidationResponse → lib/types/deployments.ts

Server and client now import from @mini-infra/types. Re-exports are added
in the client hooks for backwards compatibility with existing consumers.
